readability refactor; more interfaces prepping for upcoming unit tests
